eclipselink replace hibernate, is working.

+++ sinq-jpa/src/main/scala/io/sinq/func/Count.scala
@@ -1,9 +1,7 @@
 package io.sinq.func
 
-import java.math.BigInteger
-
 import io.sinq.provider.Column
 
-case class Count[T](val col: Column[T]) extends MethodColumn[BigInteger] {
+case class Count[T](val col: Column[T]) extends MethodColumn[Long] {
   override def identifier(): String = "count"
 }
